requirements.txt
abstract_singleton==1.0.1
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
async-generator==1.10
async-lru==2.0.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
auto_gpt_plugin_template==0.0.3
autoflake==2.1.1
autograd==1.5
autoray==0.6.3
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.39.1
black==23.3.0
bleach==5.0.1
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
confection==0.0.4
contourpy==1.0.7
controlnet-aux==0.0.5
coverage==7.2.6
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
datasets==2.13.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.16.1
dill==0.3.6
distlib==0.3.6
distro==1.8.0
dnspython==2.3.0
docker==6.1.2
docutils==0.20.1
duckduckgo-search==3.4.1
einops==0.6.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
entrypoints==0.4
evaluate==0.4.0
exceptiongroup==1.1.1
execnet==1.9.0
executing==1.0.0
fastjsonschema==2.16.1
filelock==3.12.0
flake8==6.0.0
fonttools==4.39.4
fqdn==1.5.1
frechetdist==0.6
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
future==0.18.3
gensim==4.3.1
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gTTS==2.3.1
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.16.4
identify==2.5.24
idna==3.4
imageio==2.30.0
importlib-metadata==6.6.0
iniconfig==2.0.0
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
isort==5.12.0
jaraco.classes==3.3.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonlines==3.1.0
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keyring==24.3.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
llvmlite==0.40.1
loguru==0.7.0
loralib==0.1.1
lxml==4.9.2
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mistune==2.0.4
mkdocs==1.4.3
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==2.8.8
nh3==0.2.15
nltk==3.8.1
nodeenv==1.8.0
notebook==6.4.12
notebook_shim==0.2.3
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.27.2
openapi-python-client==0.13.4
opencv-python==4.7.0.72
orjson==3.8.10
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pathy==0.10.1
pbr==5.11.1
peft @ git+https://github.com/huggingface/peft.git@6c44096c7b8d55a2ecf24be9bc68393467e1584a
PennyLane==0.31.0
PennyLane-Lightning==0.31.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
pkginfo==1.9.6
platformdirs==3.2.0
playsound==1.2.2
pluggy==1.0.0
portalocker==2.7.0
pre-commit==3.3.2
preshed==3.0.8
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==4.23.2
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.8
pyflakes==3.0.1
Pygments==2.13.0
pymdown-extensions==10.0.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyproject==1.3.1
pyrsistent==0.18.1
pysmiles==1.0.2
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.1.0
pytest-integration==0.2.3
pytest-mock==3.10.0
pytest-recording==0.12.2
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytorchcv==0.0.67
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==24.0.0
rdkit==2023.3.2
readability-lxml==0.8.1
readme-renderer==42.0
redis==4.5.5
referencing==0.30.2
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.9.2
rsa==4.9
rustworkx==0.13.0
sacrebleu==2.3.1
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.0
selenium==4.1.4
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.0.post1
six==1.16.0
sklearn==0.0.post5
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlitedict==2.1.0
srsly==2.4.6
stack-data==0.5.0
sympy==1.12
tabulate==0.9.0
terminado==0.15.0
thinc==8.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.3.3
timm==0.9.2
tinycss2==1.1.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torch-geometric==2.3.1
torch-scatter==2.1.1
torch-sparse==0.6.17
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers.git@450a181d8b963b4e896be4aac701815aa554a6bb
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.14.0
twine==4.0.2
typer==0.7.0
typing_extensions==4.6.2
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
urllib3-secure-extra==0.1.0
urlopen==1.0.0
vcrpy==4.3.1
virtualenv==20.22.0
vk==3.0
warmup-scheduler @ git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git@6b5e8953a80aef5b324104dc0c2e9b8c34d622bd
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.5
webcolors==1.13
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.5.2
wget==3.2
widgetsnbextension==4.0.8
wordcloud==1.9.2
wrapt==1.15.0
wsproto==1.2.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
